+++ src/xmat-models/xmat-lib.model.ts
@@ -1,0 +1,9 @@
+import {XmatGenericObject} from "./xmat-generic-object";
+
+export interface XmatLibStruct {
+    restTimes: XmatGenericObject<number>;
+}
+
+export const XmatLib: XmatLibStruct = {
+    restTimes: {}
+};